Proyecto ordenado i las direcciones a las pantallas
primera pantalla la de bienvenida luego boton dirige a registrar i el siguiente a login i un boton en cada vista para volver atras si se equiboca el usuario
