+++ app/src/main/java/cat/udl/tidic/amb/tournmaster/Register.java
@@ -1,14 +1,31 @@
 package cat.udl.tidic.amb.tournmaster;
 
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.os.Bundle;
-
 public class Register extends AppCompatActivity {
+    private Button volver;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_register);
+
+        Intent intent = getIntent();
+
+        volver = findViewById(R.id.btn_volver);
+
+
+    }
+
+    public void anterior(View view){
+
+        Intent anterior = new Intent(Register.this, MainActivity.class);
+        startActivity(anterior);
+
     }
 }
