+++ app/src/main/res/layout/activity_main.xml
@@ -12,6 +12,7 @@
         android:layout_width="225dp"
         android:layout_height="52dp"
         android:background="@drawable/borde_redondo"
+        android:onClick="login"
         android:text="Iniciar sesión"
         android:textColor="#FFFFFF"
         app:layout_constraintBottom_toBottomOf="parent"
@@ -26,6 +27,7 @@
         android:layout_height="81dp"
         android:layout_marginTop="264dp"
         android:background="@drawable/borde_redondo"
+        android:onClick="register"
         android:text="Registrarse"
         android:textColor="#FFFFFF"
         app:layout_constraintEnd_toEndOf="parent"
