+++ app/src/main/res/layout/activity_register.xml
@@ -243,4 +243,19 @@
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintVertical_bias="0.977" />
 
+    <Button
+        android:id="@+id/btn_volver"
+        android:layout_width="88dp"
+        android:layout_height="39dp"
+        android:background="#00897B"
+        android:drawableRight="@android:drawable/ic_menu_revert"
+        android:onClick="anterior"
+        android:text="Volver"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.021"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="0.096" />
+
 </androidx.constraintlayout.widget.ConstraintLayout>
