+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,9 @@
 <resources>
     <string name="app_name">TournMaster</string>
+    <string name="login">Login!</string>
+    <string name="username">username</string>
+    <string name="iniciar_sessi">Iniciat Sessió</string>
+    <string name="Error_Login">Error amb l usuari o contrasenya</string>
+    <string name="Tittle">Has Iniciat sessio!</string>
+    <string name="Error_Conex">Hagut un error de conexió</string>
 </resources>
